reverse commit

+++ subprojects/courses/src/main/java/nl/tudelft/sem/courses/controller/CourseController.java
@@ -4,6 +4,8 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import javax.servlet.http.HttpServletRequest;
+
+import nl.tudelft.sem.courses.util.Validate;
 import nl.tudelft.sem.shared.entity.AddCourse;
 import nl.tudelft.sem.shared.entity.BareCourse;
 import nl.tudelft.sem.courses.entity.Course;
@@ -26,7 +28,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController // This means that this class is a RestController
 @RequestMapping(path = "/course") // URL's start with /course (after Application path)
-public class CourseController extends Controller {
+public class CourseController {
 
     @Autowired
     private transient CourseRepository courseRepository;
@@ -34,22 +36,40 @@ public class CourseController extends Controller {
     @Autowired
     private transient EnrollmentRepository enrollmentRepository;
 
+    @Autowired
+    private transient Validate validate;
+
     private transient String teacherRole = "teacher";
+    protected transient String noAccessMessage = "You are not allowed to view this page. Please contact administrator.";
+
 
     /**
      * Instantiates repository needed.
      */
     public CourseController(CourseRepository courseRepository,
-                            EnrollmentRepository enrollmentRepository) {
+                            EnrollmentRepository enrollmentRepository, Validate validate) {
         this.courseRepository = courseRepository;
         this.enrollmentRepository = enrollmentRepository;
+        this.validate = validate;
+    }
+
+    public Validate getValidate() {
+        return validate;
+    }
+
+    public CourseRepository getCourseRepository() {
+        return courseRepository;
+    }
+
+    public EnrollmentRepository getEnrollmentRepository() {
+        return enrollmentRepository;
     }
 
     /**
-     * Get endpoint to retrieve all courses.
-     *
-     * @return A list of {@link BareCourse}s
-     */
+         * Get endpoint to retrieve all courses.
+         *
+         * @return A list of {@link BareCourse}s
+         */
     @GetMapping("/getAllCourses")
     @ResponseBody
     public ResponseEntity<?> getAllCourses() {
@@ -85,8 +105,8 @@ public class CourseController extends Controller {
             HttpServletRequest request, @RequestBody AddCourse addCourse)
             throws IOException, InterruptedException, JSONException {
 
-        String validation = validateRole(request, teacherRole);
-        if (validation.equals(new Message(noAccessMessage))) {
+        String validation = validate.validateRole(request, teacherRole);
+        if (validation.equals(noAccessMessage)) {
             return ResponseEntity.ok(new Message(noAccessMessage));
         }
 
@@ -117,20 +137,34 @@ public class CourseController extends Controller {
      * Deletes a new course with provided parameters.
      */
     @DeleteMapping(path = "/deleteCourse") // Map ONLY POST Requests
-    public ResponseEntity<?> deleteCourse(@RequestParam String courseId) throws JSONException {
-        Course r = courseRepository.findByCourseId(courseId);
-        if (r == null) {
-            return ResponseEntity.notFound().build();
+    public ResponseEntity<?> deleteCourse(HttpServletRequest request, @RequestParam String courseId) throws JSONException, IOException, InterruptedException {
+        String validation = validate.validateRole(request, teacherRole);
+        if (validation.equals(noAccessMessage)) {
+            return ResponseEntity.ok(new Message(noAccessMessage));
         }
-        courseRepository.delete(r);
-        return ResponseEntity.ok(new Message("Course deleted."));
+
+            Course r = courseRepository.findByCourseId(courseId);
+            if (r == null) {
+                return ResponseEntity.notFound().build();
+            }
+            courseRepository.delete(r);
+            return ResponseEntity.ok(new Message("Course deleted."));
     }
 
     /**
      * Returns a course with provided parameters.
+     *
+     *
+     * @param courseId the Id of the course to delete
      */
     @GetMapping(path = "/getCourse") // Map ONLY POST Requests
-    public ResponseEntity<?> getCourse(@RequestParam String courseId) {
+    public ResponseEntity<?> getCourse(HttpServletRequest request, @RequestParam String courseId)
+            throws JSONException, IOException, InterruptedException {
+
+        String validation = validate.validateRole(request, teacherRole);
+        if (validation.equals(noAccessMessage)) {
+            return ResponseEntity.ok(new Message(noAccessMessage));
+        }
         Course r = courseRepository.findByCourseId(courseId);
         if (r == null) {
             return ResponseEntity.notFound().build();
